generateTokenModule
