+++ src/routers/auth.routers.js
@@ -1,17 +1,12 @@
 import express from 'express';
 import bcrypt from 'bcrypt';
-import jwt from 'jsonwebtoken';
 
 import { prisma } from '../utils/prisma.util.js';
 import { signInValidator } from '../middlewares/validators/sign-in.validator.middleware.js';
 import { refreshTokenValidator } from '../middlewares/require-refresh-token.middleware.js';
 import { HTTP_STATUS } from '../constants/http-status.constant.js';
 import { MESSAGES } from '../constants/message.constant.js';
-import {
-  ACCESS_TOKEN_EXPIRED_IN,
-  REFRESH_TOKEN_EXPIRED_IN,
-  HASH_SALT,
-} from '../constants/auth.constant.js';
+import { generateToken } from '../utils/generateToken.util.js';
 
 const router = express.Router();
 
@@ -29,35 +24,12 @@ router.post('/sign-in', signInValidator, async (req, res, next) => {
         .json({ status: HTTP_STATUS.UNAUTHORIZED, message: MESSAGES.AUTH.COMMON.UNAUTHORIZED });
     }
     const payload = { id: user.id };
-    const generateToken = async (payload) => {
-      const userId = payload.id;
-      // accessToken 생성
-      const accessToken = jwt.sign(payload, process.env.ACCESS_TOKEN_SECRET_KEY, {
-        expiresIn: ACCESS_TOKEN_EXPIRED_IN,
-      });
-
-      // refreshToken 생성
-      const refreshToken = jwt.sign(payload, process.env.REFRESH_TOKEN_SECRET_KEY, {
-        expiresIn: REFRESH_TOKEN_EXPIRED_IN,
-      });
-      // refreshToken bcrypt로 해쉬하기
-      // salt round 값 환경변수, 상수로 변경하기
-      const hashedRefreshToken = await bcrypt.hash(refreshToken, HASH_SALT);
-
-      // 서버에 토큰 저장
-      await prisma.refreshToken.upsert({
-        where: { userId },
-        update: { refreshToken: hashedRefreshToken },
-        create: { userId, refreshToken: hashedRefreshToken },
-      });
-      return payload;
-    };
-    generateToken(payload);
+    const data = await generateToken(payload);
 
     return res.status(HTTP_STATUS.OK).json({
       status: HTTP_STATUS.OK,
       message: '로그인에 성공하였습니다',
-      data: { accessToken, refreshToken },
+      data,
     });
   } catch (err) {
     next(err);
@@ -68,27 +40,13 @@ router.post('/sign-in', signInValidator, async (req, res, next) => {
 router.post('/refresh', refreshTokenValidator, async (req, res, next) => {
   try {
     const user = req.user;
-    // accessToken 재발급
-    const accessToken = jwt.sign({ id: user.id }, process.env.ACCESS_TOKEN_SECRET_KEY, {
-      expiresIn: ACCESS_TOKEN_EXPIRED_IN,
-    });
-    // refreshToken 재발급
-    const refreshToken = jwt.sign({ id: user.id }, process.env.REFRESH_TOKEN_SECRET_KEY, {
-      expiresIn: REFRESH_TOKEN_EXPIRED_IN,
-    });
-    // refreshToken bcrypt로 해쉬
-    // salt round 값 환경변수, 상수로 변경
-    const hashedRefreshToken = await bcrypt.hash(refreshToken, HASH_SALT);
-    // 서버에 토큰 저장
-    await prisma.refreshToken.upsert({
-      where: { userId: user.id },
-      update: { refreshToken: hashedRefreshToken },
-      create: { userId: user.id, refreshToken: hashedRefreshToken },
-    });
+    const payload = { id: user.id };
+    const data = await generateToken(payload);
+
     return res.status(HTTP_STATUS.OK).json({
       status: HTTP_STATUS.OK,
       message: '토큰 재발급에 성공하였습니다',
-      data: { accessToken, refreshToken },
+      data,
     });
   } catch (err) {
     next(err);
